Big refactor to import rigif body logic - now handles presence/abscence of frames and headers gracefully and invisibly to user.
Added warning if user tries to use a combined file of RBTs (not sure about the best way to handle a combined RBT with no headers (which users shouldn't be making in the first place, but could).

added 'undo' functionality to camera creation and data import - thanks to Pasha van Bijlert for that one.